stat_private_mount_ok_validator: allow S_IWGRP sticky directories
This is related to systems with mount namespace isolation implemented
and enabled.

Before this change, for a directory path to be a valid mountpoint
candidate, each of its components inside chroot had to be owned by
the caller or the first pseudouser, belong to the group of the first
pseudouser, be group writable only when its sticky bit is set,
and not be writable by others.

After this change, a path component with ownership described above
and its sticky bit set is also allowed to be writable by others.